Add files via upload
Problem description (data not included): A binding isotherm is obtained for a protein-ligand interaction under conditions of ligand excess and is shown in the following table. Do these binding data indicate the presence of cooperativity in the interaction? (Answer both qualitatively and quantitatively.) If possible, estimate the apparent valency of the interaction.