Add files via upload
Problem description: 
There are multiple approaches to developing enzyme inhibitors, including: i) competitive inhibition, ii) non-competitive inhibition and iii) uncompetitive inhibition. In this problem, we will derive the equations for uncompetitive inhibition, and then compare the effects of different types of inhibitors to the case where there is iv) no inhibitor.
a) Uncompetitive inhibition occurs when the inhibitor binds to, and inhibits, the enzyme- substrate complex. Using the following schematic, provide the differential equations describing the temporal dynamics of this system.
b) From these equations, derive the Michaelis-Menten equation for the presence of uncompetitive inhibitor. State assumptions.
c) Using the Michaelis-Menten form of the equations for competitive, non-competitive, and uncompetitive inhibition, how much will the enzymatic rate be decreased by treatment with 500 nM inhibitor using the following parameters:
Enzyme - substrate rate constants: k1 = 2x105 M-1 s-1; k-1 = 5x10-4 s-1 Catalysis rate constant: kcat = 0.2 s-1
Enzyme - inhibitor rate constants: k2 = 1x105 M-1 s-1 , k−2 = 1x10-3 s-1 Initial concentrations: [E]0 = 10 nM, [S]0 = 100 nM.
 
Problem 2.
Using the same system from Question 1, the goal of this question is to interrogate the effect of the Michaelis-Menten assumption (QSSA) on the accuracy of your answers from question 1.
a) Without assuming steady-state approximation (SSA), write down the differential equations that describe how the species evolve over time. Use MATLAB to numerically integrate these equations and plot the concentration of the product as a function of time.
Do this for all four cases from problem 1 i)-iv). Put all the plots on one graph for comparison. b) Now compare your answer to a) with the steady-state approximation (SSA).
Assume steady-state assumption for all enzyme/substrate/inhibitor complexes. Write down the differential equation that describes the concentration of the product over time. Use MATLAB to numerically integrate this equation, and plot the concentration of the product as a function of time.
Do this for all four cases from problem 1 i)-iv). Put all the plots on one graph for comparison. Then plot the percentage error that the SSA introduces to the actual product concentration as a function of time for each of the four cases.
c) In a) and b) we were interested in how [P] evolves over time. Now let’s look at the dynamics of [ES] instead.
For each of the four cases i)-iv), plot the concentration of enzyme/substrate complex alone just for the early time period when it reaches its maximum value. Plot the percentage error that SSA introduces in the value of the enzyme/substrate complex as a function of time.
d) Compare the errors in P and ES – is SSA justifiable for the given set of parameters? Discuss.

Problem 3.
Crosslinking. Consider a bivalent antibody binding at equilibrium (monovalent Kd1=1nM) to a
cell-surface antigen (105 antigen molecules per cell; cell diameter, 10 microns). Plot the number of antibody molecules bound to the cell by only one binding site as a function of bulk antibody concentration in the range 0.001– 1,000 nM. Plot the number of antibody molecules bound by two binding sites on the same plot. If the bivalently bound antibody crosslinks the antigen and sends an intracellular signal, at what bulk concentration of the antibody will this signal be maximal?