Add files via upload
Problem description:
Question 1
Here we will explore the effect of receptor recycling and complex degradation on numbers of receptors and complexes on the cell surface. Following the derivation in class (L8, Sept 24), consider a system with a receptor (Rs) expressed on the cell surface with a synthesis rate of Vs. Ligand (L) binds with an on rate kf and an off rate kr to form complex (Cs) on the cell surface. Receptor is internalized (rate = ker) to form an internal pool (Ri) and recycles (rate = krec). Complex is internalized (rate = kec) to form internal pool (Ci) and recycles (rate = krec). Internalized receptors and complexes are degraded at a common rate of kdeg.
A. Provide the differential equations describing your system.
B. In the absence of ligand, derive an equation for the number of receptors on the cell surface at steady state. Account for recycling in your system.
C. Derive an equation for the number of complexes on the cell surface in the presence of ligand, again accounting for recycling in your system.
D. As discussed in class, internalized receptors likely have a higher recycling rate compared to complexes, and complexes likely degrade much faster than receptors. Modify your differential equations to include this additional complexity in your model. Specifically, use krec1 for receptor recycling, krec2 for complex recycling, kdeg1 for receptor degradation and kdeg2 for complex degradation.
E. Implement this set of equations from D in MATLAB/Python and quantify the number of receptors and complexes on the cell surface at steady state in the presence of ligand using the following parameters:
kf = 105 L M-1 sec-1 kr = 10-4 sec-1
Vs = 5x10-10 M sec-1 L = 10 nM
ker = 0.1 sec-1
kec = 1 sec-1
Rs (time=0)=10000 receptors/cell rcell = 100 um
krec1 = krec2 = 0.5 sec-1 kdeg1 = kdeg2 = 0.5 sec-1
F. Accounting for the real-world changes, how does the steady state number of receptors and complexes change as krec1 increases to 0.9 sec-1 and krec2 decreases to 0.1 sec-1, while kdeg1
decreases to 0.1 sec-1 and kdeg2 increases to 0.9 sec-1? Comment on the behavior of both
surface receptors and complexes over time (i.e. how they increase or decrease compared to each other).
Question 2.
A change in a microbe’s environment results in a 40% increase in the transcription rate of a particular gene. Yet you observe a 25% decrease in the steady state concentration of the corresponding protein. What other process(es) could have been impacted by the environmental change to account for this result? Be quantitative in your response.
Question 3.
You add a pulse of tracer molecule at concentration [P]0 to a single tank of volume V . The concentration of tracer in the exit stream is
[𝑃]! = [𝑃]!𝑒!!"
where k is the rate constant for efflux from the tank. Now consider the volume V is distributed equally amongst N tanks in series. The rate constant for transport now scales with the number of tanks. Equation 5.47 from the reading file gives the relationship for the tracer concentration exiting the system as a function of time.
. A) Reproduce figure 5.18 for the dimensionless output concentration versus dimensionless time for N = 1, 2, 4, 8, 16, 32, and 64 tanks. Each of these curves is a “residence time distribution”. This “tanks in series” model is often used to model imperfect mixing situations along the range from perfectly mixed to a pure delay (high values of N).
.
. B) The coefficient of variance C.V. is the ratio of a distribution’s standard deviation to its mean.
C.V. provides a quantitative measure of how narrow a given distribution is. Plot C.V. for the 7 curves from part a) vs. N, on a log-log plot. What number of tanks in series is necessary to achieve a distribution with a C.V. = 0.1?