Add files via upload
Problem description (figures are missing):
Question 1:
Consider the following dynamical system inside a cell:
Where E represents the dimensionless concentration of an enzyme, T represents the dimensionless concentration of a transcription factor, and γ and K are positive constants. A small-molecule ligand added to the system at dimensionless concentration Lo (constant) increases the rate of enzyme synthesis in increases the rate of enzyme synthesis in a manner directly proportional to the ligand concentration.
A. Determine the fixed points of the system (T∗, E∗). B. If Lo > γK, sketch the phase portrait. Which fixed points are stable, and
which are unstable?
C. If Lo < γK, sketch the phase portrait. Which fixed points are stable, and which are unstable?
D. Plot T∗ versus Lo (include negative values for mathematical completeness, even though such values are of course irrelevant for biological variables and parameters). Label the stable and unstable branches on the plot. There is a bifurcation in this system, but this one entails a change in the stability of existing fixed points rather than a change in the number of fixed points. At what critical value of Lo does this bifurcation occur?
 
Question 2
As we discussed in lecture, one common example used to show network dynamics is glycosylation. For this problem, please implement the Sel’kov (1968) glycolytic oscillations model:
For parameter values a = 0.04 and b = 0.9, plot two trajectories on the x,y phase plane: one that spirals into a stable orbit, and one that spirals out to a stable orbit. On a separate plot, plot x & y vs. time for each of the trajectories.
Note that when plotting the phase-plane you’ll need to use the meshgrid(x,y) function to create an xy plane, and the quiver(x,y,u,v) function with all four arguments in order to plot the vectors with components u and v at each of these points. Hint: In a phase plane what do u and v represent?
Question 3
Let’s examine zero-order ultrasensitivity quantitatively, as in Figure 6-11.
A. Using Equation 6.39, plot the output S* vs. the input (Vmax,K)/(Vmax,P), for values of Km,K and Km,P of [8,8],[0.8,0.8], and [0.08,0.08].
B. Curve fit a Hill curve nH to each of the three curves you generated in part A, and also evaluate the ratio Rv defined by Goldbeter and Koshland (Eqn. 6.42). Plot nH vs. Km,K and Rv vs. Km,K.